#11: Refactored DEFAULT_COLORS and moved hex_to_rgb to utilities package.
