+++ pdfdesigner/design/content/__init__.py
@@ -9,15 +9,13 @@ This subpackage defines classes used to represent content in your PDF.
 
 """
 
-# from .paragraph import Paragraph
+from .paragraph import Paragraph
 # from .image import Image
 # from .table import Table
 from .stylesheet import Stylesheet, Style
 from .fonts import FontDefinition, register_font_family
-from .color import Color
 
 __all__ = [
-    'Color',
     'Paragraph',
     'Image',
     'Table',
